Handle potential issue with leftover firewalls
A new cluster can be created with the name of a deleted cluster.
Since cluster deletion does not properly clean up firewall rules
there might still be leftover firewall rules from the deleted
cluster.

This change waits specifically for the new firewall rule to appear
and restricts the IPs for all firewall rules that could potentially
apply to the cluster.